+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- use direction:rtl instead of row-reverse
+
 ## [0.3.0] - 2020-05-11
 
 ### Added